# Build MultiSegment support on top of SingleSegment within the same JsonReader type