… src/System.Text.JsonLab/System/Text/Json/Utf8JsonReader_MultiSegment.cs
… tests/Benchmarks/System.Text.JsonLab/JsonReaderPerf.cs
… tests/System.Text.JsonLab.Tests/JsonReaderTests.cs
… tests/System.Text.JsonLab.Tests/JsonTestHelper.cs