Alternative to https://github.com/dotnet/corefxlab/pull/2545
- Keep a single type
- Remove use of BufferReader and leverage SingleSegment implementation
- ~Slows down the SingleSegment case (larger struct + an extra check within the reader loop).~

~The tests are expected to fail atm (isFinalBlock logic needs to be fixed).~

``` ini

BenchmarkDotNet=v0.11.1, OS=Windows 10.0.17763
Intel Core i7-6700 CPU 3.40GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100-alpha1-009667
  [Host]     : .NET Core 2.1.3 (CoreCLR 4.6.26725.06, CoreFX 4.6.26725.05), 64bit RyuJIT
  Job-RSVPIX : .NET Core 2.1.3 (CoreCLR 4.6.26725.06, CoreFX 4.6.26725.05), 64bit RyuJIT

IterationCount=5  WarmupCount=3  

```


<details>
<summary>Old data</summary>

**Before:**

|                        Method |  TestCase | segmentSize |     Mean |     Error |    StdDev |  Gen 0 | Allocated |
|------------------------------ |---------- |------------ |---------:|----------:|----------:|-------:|----------:|
|          **MultiSegmentSequence** | **Json400KB** |        **4000** | **1.783 ms** | **0.1093 ms** | **0.0284 ms** | **1.9531** |   **14072 B** |
| MultiSegmentSequenceUsingSpan | Json400KB |        4000 | 1.355 ms | 0.0915 ms | 0.0238 ms |      - |       0 B |
|         **SingleSegmentSequence** | **Json400KB** |           **?** | **1.274 ms** | **0.0904 ms** | **0.0235 ms** |      **-** |       **0 B** |

**After:**

|                                  Method |  TestCase | segmentSize |     Mean |     Error |    StdDev |    Gen 0 |    Gen 1 |    Gen 2 | Allocated |
|---------------------------------------- |---------- |------------ |---------:|----------:|----------:|---------:|---------:|---------:|----------:|
|                    **MultiSegmentSequence** | **Json400KB** |        **4000** | **1.426 ms** | **0.3260 ms** | **0.0847 ms** |   **1.9531** |        **-** |        **-** |    **8216 B** |
|           MultiSegmentSequenceUsingSpan | Json400KB |        4000 | 1.318 ms | 0.1157 ms | 0.0300 ms |        - |        - |        - |       0 B |
| MultiSegmentSequenceUsingReaderSequence | Json400KB |        4000 | 1.445 ms | 0.2424 ms | 0.0630 ms |        - |        - |        - |       0 B |
|                   **SingleSegmentSequence** | **Json400KB** |           **?** | **1.527 ms** | **0.5084 ms** | **0.1320 ms** | **248.0469** | **248.0469** | **248.0469** | **1048600 B** |

</details>


**New Data:**

``` ini

BenchmarkDotNet=v0.11.1, OS=Windows 10.0.17763
Intel Core i7-6700 CPU 3.40GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100-alpha1-009667
  [Host]     : .NET Core 2.1.3 (CoreCLR 4.6.26725.06, CoreFX 4.6.26725.05), 64bit RyuJIT
  Job-DCWLKC : .NET Core 2.1.3 (CoreCLR 4.6.26725.06, CoreFX 4.6.26725.05), 64bit RyuJIT

IterationCount=5  WarmupCount=3  

```
|                                  Method |  TestCase | numberOfBytes | segmentSize |      Mean |     Error |    StdDev | Allocated |
|---------------------------------------- |---------- |-------------- |------------ |----------:|----------:|----------:|----------:|
|                **SingleSegmentSequenceByN** | **Json400KB** |             **1** |           **?** | **35.707 ms** | **1.8923 ms** | **0.4915 ms** |       **0 B** |
|                    **MultiSegmentSequence** | **Json400KB** |             **?** |           **1** | **54.341 ms** | **0.6563 ms** | **0.1705 ms** |       **0 B** |
|                **SingleSegmentSequenceByN** | **Json400KB** |            **10** |           **?** | **22.439 ms** | **0.8085 ms** | **0.2100 ms** |       **0 B** |
|                    **MultiSegmentSequence** | **Json400KB** |             **?** |          **10** |  **6.480 ms** | **0.6039 ms** | **0.1569 ms** |       **0 B** |
|                **SingleSegmentSequenceByN** | **Json400KB** |           **100** |           **?** |  **8.638 ms** | **0.9669 ms** | **0.2512 ms** |       **0 B** |
|                    **MultiSegmentSequence** | **Json400KB** |             **?** |         **100** |  **1.909 ms** | **0.1169 ms** | **0.0304 ms** |       **0 B** |
|                **SingleSegmentSequenceByN** | **Json400KB** |          **1000** |           **?** |  **1.314 ms** | **0.0184 ms** | **0.0048 ms** |       **0 B** |
|                    **MultiSegmentSequence** | **Json400KB** |             **?** |        **1000** |  **1.365 ms** | **0.0215 ms** | **0.0056 ms** |       **0 B** |
|           MultiSegmentSequenceUsingSpan | Json400KB |             ? |        1000 |  1.346 ms | 0.0972 ms | 0.0253 ms |       0 B |
| MultiSegmentSequenceUsingReaderSequence | Json400KB |             ? |        1000 |  1.356 ms | 0.0968 ms | 0.0251 ms |       0 B |
|                **SingleSegmentSequenceByN** | **Json400KB** |          **2000** |           **?** |  **1.273 ms** | **0.0160 ms** | **0.0041 ms** |       **0 B** |
|                    **MultiSegmentSequence** | **Json400KB** |             **?** |        **2000** |  **1.306 ms** | **0.0218 ms** | **0.0057 ms** |       **0 B** |
|           MultiSegmentSequenceUsingSpan | Json400KB |             ? |        2000 |  1.375 ms | 0.2509 ms | 0.0652 ms |       0 B |
| MultiSegmentSequenceUsingReaderSequence | Json400KB |             ? |        2000 |  1.341 ms | 0.0759 ms | 0.0197 ms |       0 B |
|                **SingleSegmentSequenceByN** | **Json400KB** |          **4000** |           **?** |  **1.286 ms** | **0.1463 ms** | **0.0380 ms** |       **0 B** |
|                    **MultiSegmentSequence** | **Json400KB** |             **?** |        **4000** |  **1.314 ms** | **0.1034 ms** | **0.0269 ms** |       **0 B** |
|           MultiSegmentSequenceUsingSpan | Json400KB |             ? |        4000 |  1.311 ms | 0.0398 ms | 0.0104 ms |       0 B |
| MultiSegmentSequenceUsingReaderSequence | Json400KB |             ? |        4000 |  1.319 ms | 0.0395 ms | 0.0103 ms |       0 B |
|                **SingleSegmentSequenceByN** | **Json400KB** |          **8000** |           **?** |  **1.275 ms** | **0.0449 ms** | **0.0117 ms** |       **0 B** |
|                    **MultiSegmentSequence** | **Json400KB** |             **?** |        **8000** |  **1.291 ms** | **0.0619 ms** | **0.0161 ms** |       **0 B** |
|           MultiSegmentSequenceUsingSpan | Json400KB |             ? |        8000 |  1.663 ms | 0.6487 ms | 0.1685 ms |       0 B |
| MultiSegmentSequenceUsingReaderSequence | Json400KB |             ? |        8000 |  1.294 ms | 0.1246 ms | 0.0324 ms |       0 B |
|                   **SingleSegmentSequence** | **Json400KB** |             **?** |           **?** |  **1.257 ms** | **0.0104 ms** | **0.0027 ms** |       **0 B** |


**Before:**

|                                             Method | IsDataCompact |   TestCase |           Mean |         Error |        StdDev | Scaled | ScaledSD |  Gen 0 | Allocated |
|--------------------------------------------------- |-------------- |----------- |---------------:|--------------:|--------------:|-------:|---------:|-------:|----------:|
|               **ReaderSystemTextJsonLabSpanEmptyLoop** |         **False** |  **BasicJson** |       **885.9 ns** |      **42.24 ns** |     **10.973 ns** |   **1.00** |     **0.00** |      **-** |       **0 B** |
| ReaderSystemTextJsonLabSingleSpanSequenceEmptyLoop |         False |  BasicJson |       884.0 ns |      47.88 ns |     12.436 ns |   1.00 |     0.02 |      - |       0 B |
|  ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |         False |  BasicJson |     1,537.6 ns |      40.11 ns |     10.419 ns |   1.74 |     0.02 | 0.0076 |      40 B |
|                                                    |               |            |                |               |               |        |          |        |           |
|               **ReaderSystemTextJsonLabSpanEmptyLoop** |         **False** | **HelloWorld** |       **127.2 ns** |      **12.36 ns** |      **3.211 ns** |   **1.00** |     **0.00** |      **-** |       **0 B** |
| ReaderSystemTextJsonLabSingleSpanSequenceEmptyLoop |         False | HelloWorld |       165.4 ns |      10.80 ns |      2.804 ns |   1.30 |     0.03 |      - |       0 B |
|  ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |         False | HelloWorld |       423.5 ns |      31.77 ns |      8.251 ns |   3.33 |     0.09 | 0.0091 |      40 B |
|                                                    |               |            |                |               |               |        |          |        |           |
|               **ReaderSystemTextJsonLabSpanEmptyLoop** |         **False** |  **Json400KB** | **1,310,308.1 ns** | **117,330.11 ns** | **30,476.060 ns** |   **1.00** |     **0.00** |      **-** |       **0 B** |
| ReaderSystemTextJsonLabSingleSpanSequenceEmptyLoop |         False |  Json400KB | 1,281,277.1 ns |  93,075.84 ns | 24,176.102 ns |   0.98 |     0.03 |      - |       0 B |
|  ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |         False |  Json400KB | 1,712,605.5 ns | 132,579.99 ns | 34,437.159 ns |   1.31 |     0.04 |      - |       0 B |
|                                                    |               |            |                |               |               |        |          |        |           |
|               **ReaderSystemTextJsonLabSpanEmptyLoop** |         **False** |   **Json40KB** |   **127,429.5 ns** |  **11,357.45 ns** |  **2,950.056 ns** |   **1.00** |     **0.00** |      **-** |       **0 B** |
| ReaderSystemTextJsonLabSingleSpanSequenceEmptyLoop |         False |   Json40KB |   127,431.7 ns |  10,809.30 ns |  2,807.676 ns |   1.00 |     0.03 |      - |       0 B |
|  ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |         False |   Json40KB |   166,144.9 ns |  15,469.00 ns |  4,018.015 ns |   1.30 |     0.04 |      - |       0 B |
|                                                    |               |            |                |               |               |        |          |        |           |
|               **ReaderSystemTextJsonLabSpanEmptyLoop** |         **False** |    **Json4KB** |    **11,204.8 ns** |   **2,434.23 ns** |    **632.283 ns** |   **1.00** |     **0.00** |      **-** |       **0 B** |
| ReaderSystemTextJsonLabSingleSpanSequenceEmptyLoop |         False |    Json4KB |    10,962.0 ns |   1,827.67 ns |    474.730 ns |   0.98 |     0.06 |      - |       0 B |
|  ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |         False |    Json4KB |    15,264.9 ns |   1,336.25 ns |    347.086 ns |   1.37 |     0.07 | 0.0916 |     416 B |
|                                                    |               |            |                |               |               |        |          |        |           |
|               **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |  **BasicJson** |       **633.4 ns** |      **70.55 ns** |     **18.324 ns** |   **1.00** |     **0.00** |      **-** |       **0 B** |
| ReaderSystemTextJsonLabSingleSpanSequenceEmptyLoop |          True |  BasicJson |       686.6 ns |      52.62 ns |     13.669 ns |   1.08 |     0.03 |      - |       0 B |
|  ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True |  BasicJson |     1,150.8 ns |      81.08 ns |     21.060 ns |   1.82 |     0.06 | 0.0076 |      40 B |
|                                                    |               |            |                |               |               |        |          |        |           |
|               **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** | **HelloWorld** |       **114.0 ns** |      **15.04 ns** |      **3.905 ns** |   **1.00** |     **0.00** |      **-** |       **0 B** |
| ReaderSystemTextJsonLabSingleSpanSequenceEmptyLoop |          True | HelloWorld |       166.8 ns |      28.71 ns |      7.457 ns |   1.47 |     0.07 |      - |       0 B |
|  ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True | HelloWorld |       392.7 ns |      23.04 ns |      5.984 ns |   3.45 |     0.11 | 0.0091 |      40 B |
|                                                    |               |            |                |               |               |        |          |        |           |
|               **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |  **Json400KB** |   **984,150.1 ns** |  **45,619.57 ns** | **11,849.513 ns** |   **1.00** |     **0.00** |      **-** |       **0 B** |
| ReaderSystemTextJsonLabSingleSpanSequenceEmptyLoop |          True |  Json400KB |   946,276.9 ns |  36,316.10 ns |  9,432.973 ns |   0.96 |     0.01 |      - |       0 B |
|  ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True |  Json400KB | 1,423,760.6 ns | 159,020.98 ns | 41,305.109 ns |   1.45 |     0.04 |      - |     440 B |
|                                                    |               |            |                |               |               |        |          |        |           |
|               **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |   **Json40KB** |    **91,387.8 ns** |   **4,907.74 ns** |  **1,274.766 ns** |   **1.00** |     **0.00** |      **-** |       **0 B** |
| ReaderSystemTextJsonLabSingleSpanSequenceEmptyLoop |          True |   Json40KB |    93,097.4 ns |  11,802.94 ns |  3,065.770 ns |   1.02 |     0.03 |      - |       0 B |
|  ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True |   Json40KB |   126,884.5 ns |  11,437.14 ns |  2,970.754 ns |   1.39 |     0.03 |      - |      40 B |
|                                                    |               |            |                |               |               |        |          |        |           |
|               **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |    **Json4KB** |     **7,309.0 ns** |   **1,065.77 ns** |    **276.830 ns** |   **1.00** |     **0.00** |      **-** |       **0 B** |
| ReaderSystemTextJsonLabSingleSpanSequenceEmptyLoop |          True |    Json4KB |     7,378.3 ns |     765.29 ns |    198.782 ns |   1.01 |     0.04 |      - |       0 B |
|  ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True |    Json4KB |    10,573.3 ns |   1,075.04 ns |    279.239 ns |   1.45 |     0.06 | 0.0916 |     416 B |

**After:**

|                                             Method | IsDataCompact |   TestCase |           Mean |          Error |         StdDev | Scaled | ScaledSD | Allocated |
|--------------------------------------------------- |-------------- |----------- |---------------:|---------------:|---------------:|-------:|---------:|----------:|
|               **ReaderSystemTextJsonLabSpanEmptyLoop** |         **False** |  **BasicJson** |       **917.5 ns** |      **14.786 ns** |       **3.841 ns** |   **1.00** |     **0.00** |       **0 B** |
| ReaderSystemTextJsonLabSingleSpanSequenceEmptyLoop |         False |  BasicJson |       954.0 ns |      24.037 ns |       6.244 ns |   1.04 |     0.01 |       0 B |
|  ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |         False |  BasicJson |     1,166.1 ns |      50.829 ns |      13.203 ns |   1.27 |     0.01 |       0 B |
|                                                    |               |            |                |                |                |        |          |           |
|               **ReaderSystemTextJsonLabSpanEmptyLoop** |         **False** | **HelloWorld** |       **127.1 ns** |       **9.686 ns** |       **2.516 ns** |   **1.00** |     **0.00** |       **0 B** |
| ReaderSystemTextJsonLabSingleSpanSequenceEmptyLoop |         False | HelloWorld |       169.4 ns |       4.073 ns |       1.058 ns |   1.33 |     0.02 |       0 B |
|  ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |         False | HelloWorld |       367.2 ns |      21.605 ns |       5.612 ns |   2.89 |     0.06 |       0 B |
|                                                    |               |            |                |                |                |        |          |           |
|               **ReaderSystemTextJsonLabSpanEmptyLoop** |         **False** |  **Json400KB** | **1,320,068.7 ns** |  **73,092.412 ns** |  **18,985.482 ns** |   **1.00** |     **0.00** |       **0 B** |
| ReaderSystemTextJsonLabSingleSpanSequenceEmptyLoop |         False |  Json400KB | 1,304,512.7 ns |  67,587.801 ns |  17,555.680 ns |   0.99 |     0.02 |       0 B |
|  ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |         False |  Json400KB | 1,283,868.2 ns |  70,235.540 ns |  18,243.420 ns |   0.97 |     0.02 |       0 B |
|                                                    |               |            |                |                |                |        |          |           |
|               **ReaderSystemTextJsonLabSpanEmptyLoop** |         **False** |   **Json40KB** |   **124,922.8 ns** |   **4,451.987 ns** |   **1,156.387 ns** |   **1.00** |     **0.00** |       **0 B** |
| ReaderSystemTextJsonLabSingleSpanSequenceEmptyLoop |         False |   Json40KB |   127,242.1 ns |   1,487.407 ns |     386.348 ns |   1.02 |     0.01 |       0 B |
|  ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |         False |   Json40KB |   122,071.8 ns |   6,482.110 ns |   1,683.704 ns |   0.98 |     0.01 |       0 B |
|                                                    |               |            |                |                |                |        |          |           |
|               **ReaderSystemTextJsonLabSpanEmptyLoop** |         **False** |    **Json4KB** |    **10,810.6 ns** |     **715.594 ns** |     **185.873 ns** |   **1.00** |     **0.00** |       **0 B** |
| ReaderSystemTextJsonLabSingleSpanSequenceEmptyLoop |         False |    Json4KB |    11,130.3 ns |     753.059 ns |     195.604 ns |   1.03 |     0.02 |       0 B |
|  ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |         False |    Json4KB |    10,835.3 ns |   1,476.160 ns |     383.427 ns |   1.00 |     0.04 |       0 B |
|                                                    |               |            |                |                |                |        |          |           |
|               **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |  **BasicJson** |       **655.5 ns** |      **88.668 ns** |      **23.031 ns** |   **1.00** |     **0.00** |       **0 B** |
| ReaderSystemTextJsonLabSingleSpanSequenceEmptyLoop |          True |  BasicJson |       731.6 ns |     106.948 ns |      27.779 ns |   1.12 |     0.05 |       0 B |
|  ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True |  BasicJson |       940.4 ns |      99.205 ns |      25.768 ns |   1.44 |     0.06 |       0 B |
|                                                    |               |            |                |                |                |        |          |           |
|               **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** | **HelloWorld** |       **115.4 ns** |       **5.690 ns** |       **1.478 ns** |   **1.00** |     **0.00** |       **0 B** |
| ReaderSystemTextJsonLabSingleSpanSequenceEmptyLoop |          True | HelloWorld |       162.3 ns |      18.193 ns |       4.726 ns |   1.41 |     0.04 |       0 B |
|  ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |          True | HelloWorld |       414.7 ns |     150.955 ns |      39.210 ns |   3.59 |     0.31 |       0 B |
|                                                    |               |            |                |                |                |        |          |           |
|               **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |  **Json400KB** | **1,096,775.6 ns** | **271,912.671 ns** |  **70,628.305 ns** |   **1.00** |     **0.00** |       **0 B** |
…