@ahsonkhan 2018-10-19
@JeremyKuhne, @KrzysztofCwalina, @stephentoub - notice the 30-40% cost of supporting multi-segment disappeared.

**Before:** Json4KB | 10,573.3 ns
**After:** Json4KB | 7,844.6 ns